+++ plugin/commands/ping.js
@@ -1,5 +1,4 @@
 module.exports = {
-  crewOnly: false,
   example: 'Ping',
   accept: (msg) => (msg.data.toLowerCase() === 'ping'),
   handle: (msg, settings, device) => device.sendText('Pong', msg.from, true, false),